+++ ParticlePhysics/Entities/Cloth.cpp
@@ -119,7 +119,7 @@ void Cloth::initXY(const real dimensions[], const uint subdivision[], const real
   displayVertex.copyData(&pointPosition[0][0], subdivision[0] * subdivision[1], 0, sizeof(Real3));
 
   displayElements.gen();
-  displayElements.copyData((GLuint*)&connectionElements[0], connectionElements.size());
+  displayElements.copyData((GLuint*)&connectionElements[0], (uint)connectionElements.size());
 
   displayShader.init("SolidVert.glsl", "SolidFrag.glsl");
 #endif
